requirements.txt
cloudscraper
faker
pycryptodome
mangum